fix: [PL-31996]: Fixing manual deletion flows when FF PL_USER_DELETION is ON (#45141)
* fix: [PL-31996]: Fixing manual deletion flows when FF PL_USER_DELETION_V2 is on.

* fix: [PL-31996]: Fixing manual deletion flows when FF PL_USER_DELETION_V2 is on.

* fix: [PL-31996]: Fixing manual deletion flows when FF PL_USER_DELETION_V2 is on.

* fix: [PL-31996]: Fixing manual deletion flows when FF PL_USER_DELETION_V2 is on.

* fix: [PL-31996]: Fixing manual deletion flows when FF PL_USER_DELETION_V2 is on.

* Rebase develop.

* fix: [PL-31996]: Fixing manual deletion flows when FF PL_USER_DELETION is ON

+++ 400-rest/src/main/java/software/wings/resources/UserResourceNG.java
@@ -198,7 +198,9 @@ public class UserResourceNG {
   public RestResponse<Boolean> deleteUser(
       @QueryParam("accountId") String accountId, @QueryParam("userId") String userId) {
     if (featureFlagService.isEnabled(FeatureName.PL_USER_DELETION_V2, accountId)) {
-      userServiceHelper.deleteUserFromNG(userId, accountId, NGRemoveUserFilter.ACCOUNT_LAST_ADMIN_CHECK);
+      if (userService.isUserPresent(userId) && userServiceHelper.isUserActiveInNG(userService.get(userId), accountId)) {
+        userServiceHelper.deleteUserFromNG(userId, accountId, NGRemoveUserFilter.ACCOUNT_LAST_ADMIN_CHECK);
+      }
       if (!userService.isUserPartOfAnyUserGroupInCG(userId, accountId)) {
         log.warn("User {}, is being deleted from CG, since he is not part of any user-groups in CG",
             userService.get(userId).getEmail());
